requirements.txt
python-docx>=0.8.11
pdfplumber>=0.9.0
